[Refactor] Promote send to base backend
